v2.9.0: Applied fixes & changes according to Magento's code review comments
