--- etc/events.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0"?>
-<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:Event/etc/events.xsd">
-	<event name="layout_generate_blocks_after">
-        <observer name="yotpo_yotpo_layout_generate_blocks_after" instance="Yotpo\Yotpo\Observer\RemoveBlocks" />
-    </event>
-</config>
