+++ view/frontend/templates/widget_script.phtml
@@ -7,6 +7,11 @@
     return;
 }
 ?>
+<!-- Yotpo - Add body class -->
+<script>
+    (function(){require(["jquery"],function($){$(document).ready(function() {$('body').addClass('yotpo-yotpo-is-enabled');});});})();
+</script>
+<!--/ Yotpo - Add body class -->
 <!-- Yotpo - Widget Script -->
 <script>
     (function e(){var e=document.createElement("script");e.type="text/javascript",e.async=true,e.src="<?= $block->getYotpoWidgetUrl() ?>";var t=document.getElementsByTagName("script")[0];t.parentNode.insertBefore(e,t)})();
